requirements.txt
emoji
google-api-python-client
huggingface-hub
mcp
mcp[cli]
python-dotenv
torch
transformers